feat: convert to bigdecimal for better precision

+++ app/src/main/java/com/example/trickcalculator/utils/aliases.kt
@@ -1,5 +1,7 @@
 package com.example.trickcalculator.utils
 
+import java.math.BigDecimal
+
 typealias IntList = List<Int>
-typealias OperatorFunction = (Float, Float, String) -> Float
+typealias OperatorFunction = (BigDecimal, BigDecimal, String) -> BigDecimal
 typealias StringList = List<String>
